Add session keys and increase maximum block size for RSA encryption to 4096b
SessionKeys are symmetrical and based on a SecureRandom value generated by the original initiator of the connection.
SessionKeys are later on decrypted by the receiver and regenerated with the decrypted secret, after which the new key is sent back to the initiator, which in turn verifies the secret. If the secret is still equal, the connection is maintained.

Packet encryption will now further encrypt the already encrypted data (with the private key of the sender) with the session key. Finally the encrypted byte[] is converted into a Base64 string (private key encrypted data will still be converted to ISO-8859-1.
If either the session key, or public key is out of sync, the message can not be decrypted and deserialized.

To help better understand the key exchange process, and the handling of packets, I provided an example in JMain